# uint256: optimize and add benchmark for Set