Not sure why shallow copy is slow in golang, but changing shallow copy to direct value copy boosts `Set` from 2 ns to 0.22 ns.

Running
```
go test ./...
```
returns
```
ok  	github.com/holiman/uint256	0.975s
```

## Benchmark
```
goos: linux
goarch: amd64
pkg: github.com/holiman/uint256
cpu: AMD Ryzen 7 7735H with Radeon Graphics         
                      │     old      │                 new                  │
                      │    sec/op    │    sec/op     vs base                │
Set/single/uint256-16   2.0205n ± 1%   0.2160n ± 2%  -89.31% (p=0.000 n=10)
Set/single/big-16        2.908n ± 1%    2.925n ± 2%        ~ (p=0.516 n=10)
geomean                  2.424n        0.7950n       -67.20%

                      │     old      │                 new                 │
                      │     B/op     │    B/op     vs base                 │
Set/single/uint256-16   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Set/single/big-16       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                            ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                      │     old      │                 new                 │
                      │  allocs/op   │ allocs/op   vs base                 │
Set/single/uint256-16   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
Set/single/big-16       0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                            ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```